Added resizing tool
