--- data_processing/tests/test_crop_images.py
@@ -1,21 +1,0 @@
-from pathlib import Path
-import pytest
-
-from ..crop_and_resize_images import crop_and_resize_image, crop_and_resize_images
-from ..utils import Point, Region, Resolution
-
-@pytest.mark.parametrize("test_image, test_region, expected_image,",
-    [(Path(__file__).parent / "happy_man.png", Region(Point(100, 100), Point(400, 400)), Resolution(224, 224), Path(__file__).parent / "cropped/happy_man_c.png"),])
-def test_crop_and_resize_image(test_image, test_region, test_resolution, expected_image):
-    assert crop_and_resize_image(test_image, test_region, test_resolution) == expected_image
-    assert expected_image.exists() and expected_image.is_file()
-
-@pytest.mark.parametrize("test_images, test_region, expected_images,",
-    [([Path(__file__).parent / "happy_man.png", Path(__file__).parent / "excited_woman.png"], Region(Point(100, 100), Point(500, 500)), Resolution(224, 224), [Path(__file__).parent / "cropped/happy_man_c.png", Path(__file__).parent / "cropped/excited_woman_c.png"]),])
-def test_crop_and_resize_images(test_images, test_region, test_resolution, expected_images):
-    image_paths = crop_and_resize_images(test_images, test_region, test_resolution)
-    assert len(image_paths) == len(expected_images)
-
-    for i, image_path in enumerate(image_paths):
-        assert image_path.exists() and image_path.is_file()
-        assert image_path == expected_images[i]
