+++ data_processing/tests/test_files/happy_man.png
